# Bassera-Core/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
    "/api/users/": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get All Users",
        "operationId": "get_all_users_api_users__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/UserResponse"
                  },
                  "type": "array",
                  "title": "Response Get All Users Api Users  Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Create User",
        "operationId": "create_user_api_users__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/users/{user_id}": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get User By Id",
        "operationId": "get_user_by_id_api_users__user_id__get",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Users"
        ],
        "summary": "Update User",
        "operationId": "update_user_api_users__user_id__put",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Users"
        ],
        "summary": "Delete User",
        "operationId": "delete_user_api_users__user_id__delete",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/users/{user_id}/ai-status": {
      "patch": {
        "tags": [
          "Users"
        ],
        "summary": "Update Ai Status",
        "operationId": "update_ai_status_api_users__user_id__ai_status_patch",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AIStatusUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/users/{user_id}/ai-forecast": {
      "put": {
        "tags": [
          "Users"
        ],
        "summary": "Update Ai Forecast",
        "operationId": "update_ai_forecast_api_users__user_id__ai_forecast_put",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AIForecastPayload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/bank/": {
      "post": {
        "tags": [
          "Bank"
        ],
        "summary": "Create Bank Profile",
        "operationId": "create_bank_profile_bank__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BankCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BankResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/bank/{bank_id}": {
      "get": {
        "tags": [
          "Bank"
        ],
        "summary": "Get Bank Profile",
        "operationId": "get_bank_profile_bank__bank_id__get",
        "parameters": [
          {
            "name": "bank_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Bank Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BankResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/bank/{bank_id}/transactions": {
      "post": {
        "tags": [
          "Bank"
        ],
        "summary": "Add Transaction",
        "operationId": "add_transaction_bank__bank_id__transactions_post",
        "parameters": [
          {
            "name": "bank_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Bank Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransactionCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Bank"
        ],
        "summary": "Get Transactions",
        "operationId": "get_transactions_bank__bank_id__transactions_get",
        "parameters": [
          {
            "name": "bank_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Bank Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TransactionResponse"
                  },
                  "title": "Response Get Transactions Bank  Bank Id  Transactions Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/suggestions/{user_id}/generate": {
      "post": {
        "tags": [
          "Suggestions"
        ],
        "summary": "Generate Suggestions",
        "operationId": "generate_suggestions_api_suggestions__user_id__generate_post",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuggestionsRecord"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/suggestions/{user_id}": {
      "get": {
        "tags": [
          "Suggestions"
        ],
        "summary": "Get Latest Suggestions",
        "operationId": "get_latest_suggestions_api_suggestions__user_id__get",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuggestionsRecord"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AIForecastPayload": {
        "properties": {
          "metadata": {
            "$ref": "#/components/schemas/ForecastMetadata"
          },
          "summary": {
            "$ref": "#/components/schemas/ForecastSummary"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          },
          "rules": {
            "$ref": "#/components/schemas/ForecastRules"
          },
          "daily_forecast": {
            "items": {
              "$ref": "#/components/schemas/DailyForecastItem"
            },
            "type": "array",
            "title": "Daily Forecast"
          }
        },
        "type": "object",
        "required": [
          "metadata",
          "summary",
          "warnings",
          "rules",
          "daily_forecast"
        ],
        "title": "AIForecastPayload"
      },
      "AIStatus": {
        "type": "string",
        "enum": [
          "NOT_STARTED",
          "WORKING",
          "FINISHED",
          "FAILED"
        ],
        "title": "AIStatus"
      },
      "AIStatusUpdate": {
        "properties": {
          "ai_status": {
            "$ref": "#/components/schemas/AIStatus"
          }
        },
        "type": "object",
        "required": [
          "ai_status"
        ],
        "title": "AIStatusUpdate"
      },
      "AnalysisPeriod": {
        "properties": {
          "from": {
            "type": "string",
            "title": "From"
          },
          "to": {
            "type": "string",
            "title": "To"
          },
          "total_debits": {
            "type": "number",
            "title": "Total Debits"
          },
          "total_credits": {
            "type": "number",
            "title": "Total Credits"
          },
          "net_cashflow": {
            "type": "number",
            "title": "Net Cashflow"
          },
          "currency": {
            "type": "string",
            "title": "Currency",
            "default": "EGP"
          }
        },
        "type": "object",
        "required": [
          "from",
          "to",
          "total_debits",
          "total_credits",
          "net_cashflow"
        ],
        "title": "AnalysisPeriod"
      },
      "BankCreate": {
        "properties": {
          "bank_id": {
            "type": "string",
            "title": "Bank Id"
          }
        },
        "type": "object",
        "required": [
          "bank_id"
        ],
        "title": "BankCreate"
      },
      "BankResponse": {
        "properties": {
          "_id": {
            "type": "string",
            "title": "Id"
          },
          "bank_id": {
            "type": "string",
            "title": "Bank Id"
          },
          "balance": {
            "type": "number",
            "title": "Balance",
            "default": 0.0
          },
          "transactions": {
            "items": {
              "$ref": "#/components/schemas/TransactionResponse"
            },
            "type": "array",
            "title": "Transactions",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "_id",
          "bank_id"
        ],
        "title": "BankResponse"
      },
      "DailyForecastItem": {
        "properties": {
          "date": {
            "type": "string",
            "title": "Date"
          },
          "dynamic_income": {
            "type": "number",
            "title": "Dynamic Income"
          },
          "fixed_income": {
            "type": "number",
            "title": "Fixed Income"
          },
          "total_income": {
            "type": "number",
            "title": "Total Income"
          },
          "dynamic_expense": {
            "type": "number",
            "title": "Dynamic Expense"
          },
          "fixed_expense": {
            "type": "number",
            "title": "Fixed Expense"
          },
          "total_expense": {
            "type": "number",
            "title": "Total Expense"
          },
          "net_cash_flow": {
            "type": "number",
            "title": "Net Cash Flow"
          },
          "projected_balance": {
            "type": "number",
            "title": "Projected Balance"
          }
        },
        "type": "object",
        "required": [
          "date",
          "dynamic_income",
          "fixed_income",
          "total_income",
          "dynamic_expense",
          "fixed_expense",
          "total_expense",
          "net_cash_flow",
          "projected_balance"
        ],
        "title": "DailyForecastItem"
      },
      "ForecastMetadata": {
        "properties": {
          "forecast_horizon_days": {
            "type": "integer",
            "title": "Forecast Horizon Days"
          },
          "last_transaction_date": {
            "type": "string",
            "title": "Last Transaction Date"
          },
          "forecast_start_date": {
            "type": "string",
            "title": "Forecast Start Date"
          },
          "forecast_end_date": {
            "type": "string",
            "title": "Forecast End Date"
          },
          "model_mae_egp": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Mae Egp"
          }
        },
        "type": "object",
        "required": [
          "forecast_horizon_days",
          "last_transaction_date",
          "forecast_start_date",
          "forecast_end_date"
        ],
        "title": "ForecastMetadata"
      },
...
      "ForecastRules": {
        "properties": {
          "fixed_incomes": {
            "items": {
              "$ref": "#/components/schemas/ForecastRule"
            },
            "type": "array",
            "title": "Fixed Incomes"
          },
          "fixed_expenses": {
            "items": {
              "$ref": "#/components/schemas/ForecastRule"
            },
            "type": "array",
            "title": "Fixed Expenses"
          }
        },
        "type": "object",
        "required": [
          "fixed_incomes",
          "fixed_expenses"
        ],
        "title": "ForecastRules"
      },
      "ForecastSummary": {
        "properties": {
          "starting_balance": {
            "type": "number",
            "title": "Starting Balance"
          },
          "projected_ending_balance": {
            "type": "number",
            "title": "Projected Ending Balance"
          },
          "net_cash_flow": {
            "type": "number",
            "title": "Net Cash Flow"
          },
          "total_income": {
            "type": "number",
            "title": "Total Income"
          },
          "total_expense": {
            "type": "number",
            "title": "Total Expense"
          }
        },
        "type": "object",
        "required": [
          "starting_balance",
          "projected_ending_balance",
          "net_cash_flow",
          "total_income",
          "total_expense"
        ],
        "title": "ForecastSummary"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SavingsPotential": {
        "properties": {
          "conservative_egp": {
            "type": "number",
            "title": "Conservative Egp"
          },
          "moderate_egp": {
            "type": "number",
            "title": "Moderate Egp"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          }
        },
        "type": "object",
        "required": [
          "conservative_egp",
          "moderate_egp",
          "summary"
        ],
        "title": "SavingsPotential"
      },
      "SpendingCategory": {
        "properties": {
          "category": {
            "type": "string",
            "title": "Category"
          },
          "total": {
            "type": "number",
            "title": "Total"
          },
          "percentage_of_debits": {
            "type": "number",
            "title": "Percentage Of Debits"
          },
          "transaction_count": {
            "type": "integer",
            "title": "Transaction Count"
          }
        },
        "type": "object",
        "required": [
          "category",
          "total",
          "percentage_of_debits",
          "transaction_count"
        ],
        "title": "SpendingCategory"
      },
      "Suggestion": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "priority": {
            "$ref": "#/components/schemas/Priority"
          },
          "type": {
            "$ref": "#/components/schemas/SuggestionType"
          },
          "title": {
            "type": "string",
            "maxLength": 80,
            "title": "Title"
          },
          "body": {
            "type": "string",
            "title": "Body"
          },
          "estimated_monthly_saving_egp": {
            "type": "number",
            "minimum": 0.0,
            "title": "Estimated Monthly Saving Egp"
          },
          "affected_categories": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Affected Categories"
          },
          "action_label": {
            "type": "string",
            "maxLength": 40,
            "title": "Action Label"
          }
        },
        "type": "object",
        "required": [
          "id",
          "priority",
          "type",
          "title",
          "body",
          "estimated_monthly_saving_egp",
          "affected_categories",
          "action_label"
        ],
        "title": "Suggestion"
      },
...
      "SuggestionsRecord": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "analysis_period": {
            "$ref": "#/components/schemas/AnalysisPeriod"
          },
          "spending_breakdown": {
            "items": {
              "$ref": "#/components/schemas/SpendingCategory"
            },
            "type": "array",
            "title": "Spending Breakdown"
          },
          "suggestions": {
            "items": {
              "$ref": "#/components/schemas/Suggestion"
            },
            "type": "array",
            "title": "Suggestions"
          },
          "savings_potential": {
            "$ref": "#/components/schemas/SavingsPotential"
          },
          "insight_of_the_day": {
            "$ref": "#/components/schemas/InsightOfTheDay"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "created_at",
          "analysis_period",
          "spending_breakdown",
          "suggestions",
          "savings_potential",
          "insight_of_the_day"
        ],
        "title": "SuggestionsRecord"
      },
      "TransactionCreate": {
        "properties": {
          "merchant": {
            "type": "string",
            "title": "Merchant"
          },
          "category": {
            "type": "string",
            "title": "Category"
          },
          "amount": {
            "type": "number",
            "title": "Amount"
          },
          "currency": {
            "type": "string",
            "title": "Currency",
            "default": "EGP"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          }
        },
        "type": "object",
        "required": [
          "merchant",
          "category",
          "amount",
          "type",
          "timestamp"
        ],
        "title": "TransactionCreate"
      },
      "TransactionResponse": {
        "properties": {
          "merchant": {
            "type": "string",
            "title": "Merchant"
          },
          "category": {
            "type": "string",
            "title": "Category"
          },
          "amount": {
            "type": "number",
            "title": "Amount"
          },
          "currency": {
            "type": "string",
            "title": "Currency",
            "default": "EGP"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          },
          "transaction_id": {
            "type": "string",
            "title": "Transaction Id"
          }
        },
        "type": "object",
        "required": [
          "merchant",
          "category",
          "amount",
          "type",
          "timestamp",
          "transaction_id"
        ],
        "title": "TransactionResponse"
      },
      "UserCreate": {
        "properties": {
          "fname": {
            "type": "string",
            "title": "Fname"
          },
          "lname": {
            "type": "string",
            "title": "Lname"
          },
          "bank_id": {
            "type": "string",
            "title": "Bank Id"
          }
        },
        "type": "object",
        "required": [
          "fname",
          "lname",
          "bank_id"
        ],
        "title": "UserCreate"
      },
      "UserResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "fname": {
            "type": "string",
            "title": "Fname"
          },
          "lname": {
            "type": "string",
            "title": "Lname"
          },
          "bank_id": {
            "type": "string",
            "title": "Bank Id"
          },
          "ai_status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AIStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "ai_last_run": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ai Last Run"
          },
          "ai_forecast_data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AIForecastPayload"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "fname",
          "lname",
          "bank_id"
        ],
        "title": "UserResponse"
      },
      "UserUpdate": {
        "properties": {
          "fname": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fname"
          },
          "lname": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lname"
          },
          "bank_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bank Id"
          }
        },
        "type": "object",
        "title": "UserUpdate"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}